requirements.txt
audioread==3.0.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
confuse==1.5.0
decorator==5.1.1
idna==3.7
joblib==1.4.2
lazy-loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
loguru==0.7.2
mediafile==0.8.0
msgpack==1.0.8
mutagen==1.47.0
mysql-connector-python==9.0.0
numba==0.60.0
numpy==2.0.1
packaging==24.1
platformdirs==4.2.2
PlexAPI==4.15.15
pooch==1.8.2
pycparser==2.22
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
soundfile==0.12.1
soxr==0.4.0
threadpoolctl==3.5.0
typing-extensions==4.12.2
urllib3==2.2.2
